"Actualización del menú de navegación, eliminación de todos los fragmentos."
