+++ U-Vallet/app/src/main/java/com/example/u_vallet/Activity_EditarPerfil.java
@@ -23,6 +23,8 @@ import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.util.Log;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
@@ -98,5 +100,24 @@ public class Activity_EditarPerfil extends AppCompatActivity {
                 return;
             }
         }
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu){
+        getMenuInflater().inflate(R.menu.activity__navegation, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item){
+        int itemClicked = item.getItemId();
+        if(itemClicked == R.id.menuCambiarRol){
+            Intent intent = new Intent(this, Activity_Roles.class);
+            startActivity(intent);
+        }else if (itemClicked == R.id.menuEditarPerfil){
+            Intent intent = new Intent( this, Activity_EditarPerfil.class);
+            startActivity(intent);
+        }
+        return super.onOptionsItemSelected(item);
     }
 }
