+++ U-Vallet/app/src/main/java/com/example/u_vallet/Activity_Roles.java
@@ -9,6 +9,8 @@ import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 
@@ -37,5 +39,24 @@ public class Activity_Roles extends AppCompatActivity {
                 startActivity(intentPasajero);
             }
         });
+
+    }
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu){
+        getMenuInflater().inflate(R.menu.activity__navegation, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item){
+        int itemClicked = item.getItemId();
+        if(itemClicked == R.id.menuCambiarRol){
+            Intent intent = new Intent(this, Activity_Roles.class);
+            startActivity(intent);
+        }else if (itemClicked == R.id.menuEditarPerfil){
+            Intent intent = new Intent( this, Activity_EditarPerfil.class);
+            startActivity(intent);
+        }
+        return super.onOptionsItemSelected(item);
     }
 }
