--- U-Vallet/app/src/main/java/com/example/u_vallet/ui/editar_perfil/EditarPerfilFragment.java
@@ -1,91 +1,0 @@
-package com.example.u_vallet.ui.editar_perfil;
-
-import android.Manifest;
-import android.app.Activity;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.net.Uri;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.ImageView;
-import android.widget.Toast;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
-
-import com.example.u_vallet.R;
-import com.example.u_vallet.permisosAndroid;
-
-import java.io.FileNotFoundException;
-import java.io.InputStream;
-
-public class EditarPerfilFragment extends Fragment {
-
-    int IMAGE_PICKER_REQUEST = 1;
-    int REQUEST_IMAGE_CAPTURE = 2;
-
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             ViewGroup container, Bundle savedInstanceState) {
-
-        View root = inflater.inflate(R.layout.activity_editar_perfil, container, false);
-        permisosAndroid.requestPermission((Activity) root.getContext(), Manifest.permission.CAMERA, "", REQUEST_IMAGE_CAPTURE);
-        permisosAndroid.requestPermission((Activity) root.getContext(), Manifest.permission.READ_EXTERNAL_STORAGE, "", IMAGE_PICKER_REQUEST);
-        Button botonCambiarImagen = (Button) root.findViewById(R.id.botonCambiarImagen);
-        botonCambiarImagen.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent pickImage = new Intent(Intent.ACTION_PICK);
-                pickImage.setType("image/*");
-                startActivityForResult(pickImage, IMAGE_PICKER_REQUEST);
-            }
-        });
-        return root;
-    }
-    /*@Override
-    public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        ImageView image = (ImageView) findViewById(R.id.imagenFotoPerfil);
-        switch (requestCode) {
-            case 1: {
-                if (resultCode == RESULT_OK) {
-                    try {
-                        final Uri imageUri = data.getData();
-                        final InputStream imageStream = getContentResolver().openInputStream(imageUri);
-                        final Bitmap selectedImage = BitmapFactory.decodeStream(imageStream);
-                        image.setImageBitmap(selectedImage);
-                    } catch (FileNotFoundException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
-            case 2: {
-                if (requestCode == REQUEST_IMAGE_CAPTURE && resultCode == RESULT_OK) {
-                    Bundle extras = data.getExtras();
-                    Bitmap imageBitmap = (Bitmap) extras.get("data");
-                    image.setImageBitmap(imageBitmap);
-                }
-            }
-        }
-    }
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        switch (requestCode) {
-            case 2: {
-                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    Toast.makeText(this, "Permiso Concedido", Toast.LENGTH_SHORT).show();
-                } else {
-                    Toast.makeText(this, "Permiso Denegado", Toast.LENGTH_SHORT).show();
-                }
-                return;
-            }
-        }
-    }*/
-}
