--- U-Vallet/app/src/main/java/com/example/u_vallet/ui/mis_carros/MisCarrosFragment.java
@@ -1,27 +1,0 @@
-package com.example.u_vallet.ui.mis_carros;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
-
-import com.example.u_vallet.R;
-import com.example.u_vallet.ui.slideshow.SlideshowViewModel;
-
-public class MisCarrosFragment extends Fragment {
-    private SlideshowViewModel slideshowViewModel;
-
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             ViewGroup container, Bundle savedInstanceState) {
-        slideshowViewModel =
-                new ViewModelProvider(this).get(SlideshowViewModel.class);
-        View root = inflater.inflate(R.layout.activity_mis_carros, container, false);
-        return root;
-    }
-
-}
