--- U-Vallet/app/src/main/java/com/example/u_vallet/ui/slideshow/SlideshowViewModel.java
@@ -1,19 +1,0 @@
-package com.example.u_vallet.ui.slideshow;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class SlideshowViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public SlideshowViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is slideshow fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
