+++ U-Vallet/app/src/main/res/layout/activity_login.xml
@@ -227,11 +227,11 @@
                 android:layout_alignParentLeft="true"
                 android:layout_alignParentEnd="true"
                 android:layout_alignParentRight="true"
-                android:layout_marginStart="83dp"
-                android:layout_marginLeft="83dp"
-                android:layout_marginTop="63dp"
-                android:layout_marginEnd="37dp"
-                android:layout_marginRight="37dp"
+                android:layout_marginStart="71dp"
+                android:layout_marginLeft="71dp"
+                android:layout_marginTop="62dp"
+                android:layout_marginEnd="49dp"
+                android:layout_marginRight="49dp"
                 android:background="#FFC107"
                 android:singleLine="false"
                 android:text="Ingresar con Facebook"
