+++ U-Vallet/app/src/main/res/menu/activity__navegation.xml
@@ -3,8 +3,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
         android:id="@+id/menuCambiarRol" android:title="Cambiar Rol"/>
-    <!--<item
-        android:id="@+id/menuEditarPerfil" android:title="Editar Perfil"/>-->
+    <item
+        android:id="@+id/menuEditarPerfil" android:title="Editar Perfil"/>
     <item
         android:id="@+id/menuLogOut" android:title="Cerrar Sesión"/>
