--- U-Vallet/app/src/main/res/menu/activity_main_drawer.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="navigation_view">
-
-    <group android:checkableBehavior="single">
-        <item
-            android:id="@+id/nav_rol"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="@string/menu_home" />
-        <item
-            android:id="@+id/nav_editar_perfil"
-            android:icon="@drawable/ic_menu_gallery"
-            android:title="@string/menu_editar_perfil" />
-
-    </group>
-</menu>
